+++ x/bank/keeper/msg_server.go
@@ -2,17 +2,28 @@ package keeper
 
 import (
 	"context"
-
 	"github.com/armon/go-metrics"
-
+	"cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
+	gov "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
 type msgServer struct {
 	Keeper
+}
+
+// UpdateDenomMetadata updates the denom metadata if the message is signed by gov module account
+func (k msgServer) UpdateDenomMetadata(goCtx context.Context, msg *types.MsgUpdateDenomMetadata) (*types.MsgUpdateDenomMetadataResponse, error) {
+	if k.GetAuthority() != msg.FromAddress {
+		return nil, errors.Wrapf(gov.ErrInvalidSigner, "expected %s got %s", k.GetAuthority(), msg.FromAddress)
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	k.Keeper.SetDenomMetaData(ctx, *msg.Metadata)
+	return &types.MsgUpdateDenomMetadataResponse{}, nil
 }
 
 // NewMsgServerImpl returns an implementation of the bank MsgServer interface
@@ -102,3 +113,4 @@ func (k msgServer) MultiSend(goCtx context.Context, msg *types.MsgMultiSend) (*t
 
 	return &types.MsgMultiSendResponse{}, nil
 }
+