+++ x/bank/types/msgs_test.go
@@ -275,3 +275,35 @@ func TestMsgSendGetSigners(t *testing.T) {
 	require.Equal(t, 1, len(res))
 	require.True(t, from.Equals(res[0]))
 }
+
+func TestUpdateDenomMetadataGetSignBytes(t *testing.T) {
+	//from := sdk.AccAddress("input")
+	//coins := sdk.NewCoins(sdk.NewInt64Coin("atom", 10))
+	msg := MsgUpdateDenomMetadata{
+		Title: "title",
+		Description: "description",
+		Metadata: &Metadata {
+			Name:        "Cosmos Hub Atom",
+			Symbol:      "ATOM",
+			Description: "The native staking token of the Cosmos Hub.",
+			DenomUnits: []*DenomUnit{
+				{"uatom", uint32(0), []string{"microatom"}},
+			},
+		},
+	}
+	res := msg.GetSignBytes()
+
+	expected := `{"type":"cosmos-sdk/MsgUpdateDenomMetadata","value":{"description":"description","metadata":{"denom_units":[{"aliases":["microatom"],"denom":"uatom"}],"description":"The native staking token of the Cosmos Hub.","name":"Cosmos Hub Atom","symbol":"ATOM"},"title":"title"}}`
+	require.Equal(t, expected, string(res))
+}
+
+func TestUpdateDenomMetadataGetSigners(t *testing.T) {
+	from := sdk.AccAddress("input111111111111111")
+	title := "Proposal Title"
+	description := "Proposal description"
+	metadata := Metadata{}
+	msg := NewMsgUpdateDenomMetadata(from.String(), title, description, &metadata)
+	res := msg.GetSigners()
+	require.Equal(t, 1, len(res))
+	require.True(t, from.Equals(res[0]))
+}