+++ x/bank/types/tx.pb.go
@@ -199,46 +199,130 @@ func (m *MsgMultiSendResponse) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_MsgMultiSendResponse proto.InternalMessageInfo
 
+type MsgUpdateDenomMetadata struct {
+	FromAddress string    `protobuf:"bytes,1,opt,name=from_address,json=fromAddress,proto3" json:"from_address,omitempty"`
+	Title       string    `protobuf:"bytes,2,opt,name=title,proto3" json:"title,omitempty"`
+	Description string    `protobuf:"bytes,3,opt,name=description,proto3" json:"description,omitempty"`
+	Metadata    *Metadata `protobuf:"bytes,4,opt,name=metadata,proto3" json:"metadata,omitempty"`
+}
+
+func (m *MsgUpdateDenomMetadata) Reset()         { *m = MsgUpdateDenomMetadata{} }
+func (m *MsgUpdateDenomMetadata) String() string { return proto.CompactTextString(m) }
+func (*MsgUpdateDenomMetadata) ProtoMessage()    {}
+func (*MsgUpdateDenomMetadata) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1d8cb1613481f5b7, []int{4}
+}
+func (m *MsgUpdateDenomMetadata) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *MsgUpdateDenomMetadata) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_MsgUpdateDenomMetadata.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *MsgUpdateDenomMetadata) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_MsgUpdateDenomMetadata.Merge(m, src)
+}
+func (m *MsgUpdateDenomMetadata) XXX_Size() int {
+	return m.Size()
+}
+func (m *MsgUpdateDenomMetadata) XXX_DiscardUnknown() {
+	xxx_messageInfo_MsgUpdateDenomMetadata.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_MsgUpdateDenomMetadata proto.InternalMessageInfo
+
+type MsgUpdateDenomMetadataResponse struct {
+}
+
+func (m *MsgUpdateDenomMetadataResponse) Reset()         { *m = MsgUpdateDenomMetadataResponse{} }
+func (m *MsgUpdateDenomMetadataResponse) String() string { return proto.CompactTextString(m) }
+func (*MsgUpdateDenomMetadataResponse) ProtoMessage()    {}
+func (*MsgUpdateDenomMetadataResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1d8cb1613481f5b7, []int{5}
+}
+func (m *MsgUpdateDenomMetadataResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *MsgUpdateDenomMetadataResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_MsgUpdateDenomMetadataResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *MsgUpdateDenomMetadataResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_MsgUpdateDenomMetadataResponse.Merge(m, src)
+}
+func (m *MsgUpdateDenomMetadataResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *MsgUpdateDenomMetadataResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_MsgUpdateDenomMetadataResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_MsgUpdateDenomMetadataResponse proto.InternalMessageInfo
+
 func init() {
 	proto.RegisterType((*MsgSend)(nil), "cosmos.bank.v1beta1.MsgSend")
 	proto.RegisterType((*MsgSendResponse)(nil), "cosmos.bank.v1beta1.MsgSendResponse")
 	proto.RegisterType((*MsgMultiSend)(nil), "cosmos.bank.v1beta1.MsgMultiSend")
 	proto.RegisterType((*MsgMultiSendResponse)(nil), "cosmos.bank.v1beta1.MsgMultiSendResponse")
+	proto.RegisterType((*MsgUpdateDenomMetadata)(nil), "cosmos.bank.v1beta1.MsgUpdateDenomMetadata")
+	proto.RegisterType((*MsgUpdateDenomMetadataResponse)(nil), "cosmos.bank.v1beta1.MsgUpdateDenomMetadataResponse")
 }
 
 func init() { proto.RegisterFile("cosmos/bank/v1beta1/tx.proto", fileDescriptor_1d8cb1613481f5b7) }
 
 var fileDescriptor_1d8cb1613481f5b7 = []byte{
-	// 460 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x84, 0x52, 0x3d, 0x6f, 0xd3, 0x40,
-	0x18, 0xf6, 0x35, 0x55, 0xaa, 0x5c, 0x2b, 0x55, 0x98, 0x08, 0x12, 0x53, 0x39, 0xc5, 0x62, 0x48,
-	0x91, 0x7a, 0x26, 0x65, 0x00, 0xb5, 0x13, 0xee, 0x04, 0x92, 0x85, 0xe4, 0x4e, 0xb0, 0x54, 0xfe,
-	0x38, 0x0e, 0xab, 0xf8, 0xce, 0xf2, 0x9d, 0xab, 0xb2, 0x32, 0x31, 0x32, 0x31, 0x77, 0x46, 0x0c,
-	0x0c, 0xfc, 0x88, 0x8e, 0x15, 0x13, 0x13, 0xa0, 0x64, 0x80, 0x7f, 0x01, 0xba, 0x0f, 0x9b, 0x22,
-	0xd2, 0x66, 0x3a, 0xcb, 0xcf, 0xc7, 0x3d, 0xef, 0xf3, 0x1e, 0xdc, 0x48, 0x19, 0x2f, 0x18, 0xf7,
-	0x93, 0x98, 0x1e, 0xf9, 0xc7, 0x93, 0x04, 0x8b, 0x78, 0xe2, 0x8b, 0x13, 0x54, 0x56, 0x4c, 0x30,
-	0xfb, 0xba, 0x46, 0x91, 0x44, 0x91, 0x41, 0x9d, 0x3e, 0x61, 0x84, 0x29, 0xdc, 0x97, 0x5f, 0x9a,
-	0xea, 0xb8, 0xad, 0x11, 0xc7, 0xad, 0x51, 0xca, 0x72, 0xfa, 0x1f, 0x7e, 0xe1, 0x22, 0xe5, 0xab,
-	0xf1, 0xa1, 0xc6, 0x0f, 0xb5, 0xb1, 0xb9, 0x57, 0x43, 0x37, 0x8d, 0xb4, 0xe0, 0xc4, 0x3f, 0x9e,
-	0xc8, 0x43, 0x03, 0xde, 0x6f, 0x00, 0x57, 0x42, 0x4e, 0x0e, 0x30, 0xcd, 0xec, 0x3d, 0xb8, 0xf6,
-	0xa2, 0x62, 0xc5, 0x61, 0x9c, 0x65, 0x15, 0xe6, 0x7c, 0x00, 0x36, 0xc1, 0xb8, 0x17, 0x0c, 0xbe,
-	0x7c, 0xde, 0xee, 0x1b, 0xb3, 0x47, 0x1a, 0x39, 0x10, 0x55, 0x4e, 0x49, 0xb4, 0x2a, 0xd9, 0xe6,
-	0x97, 0xfd, 0x00, 0x42, 0xc1, 0x5a, 0xe9, 0xd2, 0x02, 0x69, 0x4f, 0xb0, 0x46, 0x98, 0xc2, 0x6e,
-	0x5c, 0xb0, 0x9a, 0x8a, 0x41, 0x67, 0xb3, 0x33, 0x5e, 0xdd, 0x19, 0xa2, 0xb6, 0x31, 0x8e, 0x9b,
-	0xc6, 0xd0, 0x3e, 0xcb, 0x69, 0x70, 0xef, 0xec, 0xdb, 0xc8, 0xfa, 0xf0, 0x7d, 0x34, 0x26, 0xb9,
-	0x78, 0x59, 0x27, 0x28, 0x65, 0x85, 0x19, 0xd3, 0x1c, 0xdb, 0x3c, 0x3b, 0xf2, 0xc5, 0xeb, 0x12,
-	0x73, 0x25, 0xe0, 0x91, 0xb1, 0xde, 0x1d, 0xbe, 0x3d, 0x1d, 0x59, 0xbf, 0x4e, 0x47, 0xd6, 0x9b,
-	0x9f, 0x9f, 0xee, 0xfe, 0x33, 0xa5, 0x77, 0x0d, 0xae, 0x9b, 0x02, 0x22, 0xcc, 0x4b, 0x46, 0x39,
-	0xf6, 0xde, 0x03, 0xb8, 0x16, 0x72, 0x12, 0xd6, 0xaf, 0x44, 0xae, 0x9a, 0x79, 0x08, 0xbb, 0x39,
-	0x2d, 0x6b, 0x21, 0x3b, 0x91, 0x19, 0x1d, 0x34, 0x67, 0xab, 0xe8, 0xb1, 0xa4, 0x04, 0xcb, 0x32,
-	0x64, 0x64, 0xf8, 0xf6, 0x1e, 0x5c, 0x61, 0xb5, 0x50, 0xd2, 0x25, 0x25, 0xbd, 0x35, 0x57, 0xfa,
-	0x54, 0x71, 0x8c, 0xb6, 0x51, 0xec, 0xae, 0x37, 0x89, 0x8d, 0x9b, 0x77, 0x03, 0xf6, 0x2f, 0xe6,
-	0x6a, 0x02, 0xef, 0x7c, 0x04, 0xb0, 0x13, 0x72, 0x62, 0x3f, 0x81, 0xcb, 0x2a, 0xef, 0xc6, 0xdc,
-	0x4b, 0xcc, 0x98, 0xce, 0x9d, 0xab, 0xd0, 0xc6, 0xd3, 0x7e, 0x06, 0x7b, 0x7f, 0x0b, 0xb8, 0x7d,
-	0x99, 0xa4, 0xa5, 0x38, 0x5b, 0x0b, 0x29, 0x8d, 0x75, 0xb0, 0x7f, 0x36, 0x75, 0xc1, 0xf9, 0xd4,
-	0x05, 0x3f, 0xa6, 0x2e, 0x78, 0x37, 0x73, 0xad, 0xf3, 0x99, 0x6b, 0x7d, 0x9d, 0xb9, 0xd6, 0xf3,
-	0xad, 0x2b, 0x37, 0x7b, 0xa2, 0x9f, 0xbe, 0x5a, 0x70, 0xd2, 0x55, 0x0f, 0xf8, 0xfe, 0x9f, 0x00,
-	0x00, 0x00, 0xff, 0xff, 0xa7, 0xd3, 0x87, 0x2b, 0x7f, 0x03, 0x00, 0x00,
+	// 560 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x54, 0x3d, 0x6f, 0xd3, 0x40,
+	0x18, 0xb6, 0x9b, 0x90, 0x92, 0x37, 0x95, 0x2a, 0xdc, 0xa8, 0x24, 0xa6, 0x38, 0x21, 0x62, 0x48,
+	0x41, 0xb5, 0x49, 0x3a, 0x00, 0xed, 0x44, 0xca, 0x02, 0x92, 0x85, 0xe4, 0x8a, 0x01, 0x96, 0xca,
+	0xb1, 0x0f, 0x63, 0xb5, 0xbe, 0xb3, 0x7c, 0xe7, 0x52, 0x56, 0x26, 0x46, 0x26, 0xe6, 0x2e, 0x2c,
+	0x4c, 0x0c, 0xfc, 0x88, 0x8e, 0x15, 0x13, 0x13, 0xa0, 0x64, 0x80, 0x7f, 0x01, 0xf2, 0xdd, 0xd9,
+	0x04, 0xe1, 0xa6, 0x42, 0x4c, 0x4e, 0xfc, 0x7c, 0xbc, 0x1f, 0x8f, 0xef, 0x60, 0xcd, 0x23, 0x34,
+	0x22, 0xd4, 0x1a, 0xbb, 0x78, 0xdf, 0x3a, 0x1c, 0x8c, 0x11, 0x73, 0x07, 0x16, 0x3b, 0x32, 0xe3,
+	0x84, 0x30, 0xa2, 0xad, 0x08, 0xd4, 0xcc, 0x50, 0x53, 0xa2, 0x7a, 0x33, 0x20, 0x01, 0xe1, 0xb8,
+	0x95, 0xfd, 0x12, 0x54, 0xdd, 0x28, 0x8c, 0x28, 0x2a, 0x8c, 0x3c, 0x12, 0xe2, 0xbf, 0xf0, 0x99,
+	0x42, 0xdc, 0x57, 0xe0, 0x6d, 0x81, 0xef, 0x09, 0x63, 0x59, 0x57, 0x40, 0x97, 0xa5, 0x34, 0xa2,
+	0x81, 0x75, 0x38, 0xc8, 0x1e, 0x02, 0xe8, 0xfd, 0x54, 0x61, 0xd1, 0xa6, 0xc1, 0x2e, 0xc2, 0xbe,
+	0xb6, 0x0d, 0x4b, 0xcf, 0x12, 0x12, 0xed, 0xb9, 0xbe, 0x9f, 0x20, 0x4a, 0x5b, 0x6a, 0x57, 0xed,
+	0xd7, 0x47, 0xad, 0x4f, 0x1f, 0x37, 0x9a, 0xd2, 0xec, 0x9e, 0x40, 0x76, 0x59, 0x12, 0xe2, 0xc0,
+	0x69, 0x64, 0x6c, 0xf9, 0x4a, 0xbb, 0x0d, 0xc0, 0x48, 0x21, 0x5d, 0x38, 0x47, 0x5a, 0x67, 0x24,
+	0x17, 0x7a, 0x50, 0x73, 0x23, 0x92, 0x62, 0xd6, 0xaa, 0x74, 0x2b, 0xfd, 0xc6, 0xb0, 0x6d, 0x16,
+	0x1b, 0xa3, 0x28, 0xdf, 0x98, 0xb9, 0x43, 0x42, 0x3c, 0xba, 0x75, 0xf2, 0xa5, 0xa3, 0xbc, 0xff,
+	0xda, 0xe9, 0x07, 0x21, 0x7b, 0x9e, 0x8e, 0x4d, 0x8f, 0x44, 0x72, 0x4c, 0xf9, 0xd8, 0xa0, 0xfe,
+	0xbe, 0xc5, 0x5e, 0xc6, 0x88, 0x72, 0x01, 0x75, 0xa4, 0xf5, 0x56, 0xfb, 0xf5, 0x71, 0x47, 0xf9,
+	0x71, 0xdc, 0x51, 0x5e, 0x7d, 0xff, 0x70, 0xe3, 0x8f, 0x29, 0x7b, 0x97, 0x60, 0x59, 0x2e, 0xc0,
+	0x41, 0x34, 0x26, 0x98, 0xa2, 0xde, 0x5b, 0x15, 0x96, 0x6c, 0x1a, 0xd8, 0xe9, 0x01, 0x0b, 0xf9,
+	0x66, 0xee, 0x40, 0x2d, 0xc4, 0x71, 0xca, 0xb2, 0x9d, 0x64, 0x3d, 0xea, 0x66, 0x49, 0xaa, 0xe6,
+	0x83, 0x8c, 0x32, 0xaa, 0x66, 0x4d, 0x3a, 0x92, 0xaf, 0x6d, 0xc3, 0x22, 0x49, 0x19, 0x97, 0x2e,
+	0x70, 0xe9, 0x95, 0x52, 0xe9, 0x23, 0xce, 0x91, 0xda, 0x5c, 0xb1, 0xb5, 0x9c, 0x77, 0x2c, 0xdd,
+	0x7a, 0xab, 0xd0, 0x9c, 0xed, 0xab, 0x68, 0x78, 0xaa, 0xc2, 0xaa, 0x4d, 0x83, 0xc7, 0xb1, 0xef,
+	0x32, 0x74, 0x1f, 0x61, 0x12, 0xd9, 0x88, 0xb9, 0xbe, 0xcb, 0xdc, 0xff, 0x0b, 0xb5, 0x09, 0x17,
+	0x58, 0xc8, 0x0e, 0x90, 0xc8, 0xd3, 0x11, 0x7f, 0xb4, 0x2e, 0x34, 0x7c, 0x44, 0xbd, 0x24, 0x8c,
+	0x59, 0x48, 0x70, 0xab, 0xc2, 0xb1, 0xd9, 0x57, 0xda, 0x5d, 0xb8, 0x18, 0xc9, 0x06, 0x5a, 0xd5,
+	0xae, 0xda, 0x6f, 0x0c, 0xaf, 0x96, 0x8e, 0x9d, 0x77, 0xe9, 0x14, 0xf4, 0x79, 0x49, 0x75, 0xc1,
+	0x28, 0x1f, 0x32, 0xdf, 0xc3, 0xf0, 0xdd, 0x02, 0x54, 0x6c, 0x1a, 0x68, 0x0f, 0xa1, 0xca, 0x73,
+	0x5b, 0x2b, 0xaf, 0x2a, 0xe2, 0xd6, 0xaf, 0xcf, 0x43, 0x73, 0x4f, 0xed, 0x09, 0xd4, 0x7f, 0x7f,
+	0x08, 0xd7, 0xce, 0x92, 0x14, 0x14, 0x7d, 0xfd, 0x5c, 0x4a, 0x61, 0xfd, 0x02, 0x56, 0xca, 0x22,
+	0xbb, 0x79, 0x96, 0x43, 0x09, 0x59, 0xdf, 0xfc, 0x07, 0x72, 0x5e, 0x78, 0xb4, 0x73, 0x32, 0x31,
+	0xd4, 0xd3, 0x89, 0xa1, 0x7e, 0x9b, 0x18, 0xea, 0x9b, 0xa9, 0xa1, 0x9c, 0x4e, 0x0d, 0xe5, 0xf3,
+	0xd4, 0x50, 0x9e, 0xae, 0xcf, 0x3d, 0x5a, 0x47, 0xe2, 0xee, 0xe1, 0x27, 0x6c, 0x5c, 0xe3, 0x37,
+	0xc8, 0xe6, 0xaf, 0x00, 0x00, 0x00, 0xff, 0xff, 0x4d, 0xf3, 0xf7, 0x2a, 0x00, 0x05, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -257,6 +341,8 @@ type MsgClient interface {
 	Send(ctx context.Context, in *MsgSend, opts ...grpc.CallOption) (*MsgSendResponse, error)
 	// MultiSend defines a method for sending coins from some accounts to other accounts.
 	MultiSend(ctx context.Context, in *MsgMultiSend, opts ...grpc.CallOption) (*MsgMultiSendResponse, error)
+	// UpdateDenomMetadata defines a method for updating the fee denom metadata
+	UpdateDenomMetadata(ctx context.Context, in *MsgUpdateDenomMetadata, opts ...grpc.CallOption) (*MsgUpdateDenomMetadataResponse, error)
 }
 
 type msgClient struct {
@@ -285,12 +371,23 @@ func (c *msgClient) MultiSend(ctx context.Context, in *MsgMultiSend, opts ...grp
 	return out, nil
 }
 
+func (c *msgClient) UpdateDenomMetadata(ctx context.Context, in *MsgUpdateDenomMetadata, opts ...grpc.CallOption) (*MsgUpdateDenomMetadataResponse, error) {
+	out := new(MsgUpdateDenomMetadataResponse)
+	err := c.cc.Invoke(ctx, "/cosmos.bank.v1beta1.Msg/UpdateDenomMetadata", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // MsgServer is the server API for Msg service.
 type MsgServer interface {
 	// Send defines a method for sending coins from one account to another account.
 	Send(context.Context, *MsgSend) (*MsgSendResponse, error)
 	// MultiSend defines a method for sending coins from some accounts to other accounts.
 	MultiSend(context.Context, *MsgMultiSend) (*MsgMultiSendResponse, error)
+	// UpdateDenomMetadata defines a method for updating the fee denom metadata
+	UpdateDenomMetadata(context.Context, *MsgUpdateDenomMetadata) (*MsgUpdateDenomMetadataResponse, error)
 }
 
 // UnimplementedMsgServer can be embedded to have forward compatible implementations.
@@ -302,6 +399,9 @@ func (*UnimplementedMsgServer) Send(ctx context.Context, req *MsgSend) (*MsgSend
 }
 func (*UnimplementedMsgServer) MultiSend(ctx context.Context, req *MsgMultiSend) (*MsgMultiSendResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method MultiSend not implemented")
+}
+func (*UnimplementedMsgServer) UpdateDenomMetadata(ctx context.Context, req *MsgUpdateDenomMetadata) (*MsgUpdateDenomMetadataResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method UpdateDenomMetadata not implemented")
 }
 
 func RegisterMsgServer(s grpc1.Server, srv MsgServer) {
@@ -344,6 +444,24 @@ func _Msg_MultiSend_Handler(srv interface{}, ctx context.Context, dec func(inter
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Msg_UpdateDenomMetadata_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(MsgUpdateDenomMetadata)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(MsgServer).UpdateDenomMetadata(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/cosmos.bank.v1beta1.Msg/UpdateDenomMetadata",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(MsgServer).UpdateDenomMetadata(ctx, req.(*MsgUpdateDenomMetadata))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 var _Msg_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "cosmos.bank.v1beta1.Msg",
 	HandlerType: (*MsgServer)(nil),
@@ -355,6 +473,10 @@ var _Msg_serviceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "MultiSend",
 			Handler:    _Msg_MultiSend_Handler,
+		},
+		{
+			MethodName: "UpdateDenomMetadata",
+			Handler:    _Msg_UpdateDenomMetadata_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
@@ -509,6 +631,85 @@ func (m *MsgMultiSendResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	return len(dAtA) - i, nil
 }
 
+func (m *MsgUpdateDenomMetadata) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *MsgUpdateDenomMetadata) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *MsgUpdateDenomMetadata) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Metadata != nil {
+		{
+			size, err := m.Metadata.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintTx(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x22
+	}
+	if len(m.Description) > 0 {
+		i -= len(m.Description)
+		copy(dAtA[i:], m.Description)
+		i = encodeVarintTx(dAtA, i, uint64(len(m.Description)))
+		i--
+		dAtA[i] = 0x1a
+	}
+	if len(m.Title) > 0 {
+		i -= len(m.Title)
+		copy(dAtA[i:], m.Title)
+		i = encodeVarintTx(dAtA, i, uint64(len(m.Title)))
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.FromAddress) > 0 {
+		i -= len(m.FromAddress)
+		copy(dAtA[i:], m.FromAddress)
+		i = encodeVarintTx(dAtA, i, uint64(len(m.FromAddress)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *MsgUpdateDenomMetadataResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *MsgUpdateDenomMetadataResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *MsgUpdateDenomMetadataResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	return len(dAtA) - i, nil
+}
+
 func encodeVarintTx(dAtA []byte, offset int, v uint64) int {
 	offset -= sovTx(v)
 	base := offset
@@ -574,6 +775,40 @@ func (m *MsgMultiSend) Size() (n int) {
 }
 
 func (m *MsgMultiSendResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	return n
+}
+
+func (m *MsgUpdateDenomMetadata) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.FromAddress)
+	if l > 0 {
+		n += 1 + l + sovTx(uint64(l))
+	}
+	l = len(m.Title)
+	if l > 0 {
+		n += 1 + l + sovTx(uint64(l))
+	}
+	l = len(m.Description)
+	if l > 0 {
+		n += 1 + l + sovTx(uint64(l))
+	}
+	if m.Metadata != nil {
+		l = m.Metadata.Size()
+		n += 1 + l + sovTx(uint64(l))
+	}
+	return n
+}
+
+func (m *MsgUpdateDenomMetadataResponse) Size() (n int) {
 	if m == nil {
 		return 0
 	}
@@ -931,6 +1166,238 @@ func (m *MsgMultiSendResponse) Unmarshal(dAtA []byte) error {
 		}
 		if fieldNum <= 0 {
 			return fmt.Errorf("proto: MsgMultiSendResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		default:
+			iNdEx = preIndex
+			skippy, err := skipTx(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthTx
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *MsgUpdateDenomMetadata) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowTx
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: MsgUpdateDenomMetadata: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: MsgUpdateDenomMetadata: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field FromAddress", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.FromAddress = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Title", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Title = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Description", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Description = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 4:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Metadata", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Metadata == nil {
+				m.Metadata = &Metadata{}
+			}
+			if err := m.Metadata.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipTx(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthTx
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *MsgUpdateDenomMetadataResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowTx
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: MsgUpdateDenomMetadataResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: MsgUpdateDenomMetadataResponse: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		default: